% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRatePhysioExperiment-class.R
\name{alignStreams}
\alias{alignStreams}
\title{Align all streams to the reference rate}
\usage{
alignStreams(x)
}
\arguments{
\item{x}{A \code{MultiRatePhysioExperiment}.}
}
\value{
A single-rate \code{PhysioExperiment}.
}
\description{
Convenience wrapper for \code{resampleToCommon(x, reference_rate)} using the
master clock's reference rate.
}
\seealso{
\code{\link{resampleToCommon}}
}
