% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRatePhysioExperiment-class.R
\name{commonClock}
\alias{commonClock}
\title{The master clock}
\usage{
commonClock(x)
}
\arguments{
\item{x}{A \code{MultiRatePhysioExperiment}.}
}
\value{
The master-clock list (\code{t0}, \code{reference_rate}, \code{offsets}).
}
\description{
The master clock
}
