% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRatePhysioExperiment-class.R
\name{nStreams}
\alias{nStreams}
\title{Number of streams}
\usage{
nStreams(x)
}
\arguments{
\item{x}{A \code{MultiRatePhysioExperiment}.}
}
\value{
Integer count of streams.
}
\description{
Number of streams
}
