% Generated by physio-ecosystem/publishing/generate_core_compat.R
\docType{import}
\name{PhysioCore-reexports}
\alias{PhysioCore-reexports}
\alias{PhysioExperiment}
\alias{PhysioEvents}
\alias{MultiPhysioExperiment}
\alias{streams}
\alias{streams<-}
\alias{streamNames}
\alias{streamRates}
\alias{commonClock}
\alias{nStreams}
\alias{streamTimeIndex}
\alias{hasMeasuredTimes}
\alias{timeWindow}
\alias{resampleToCommon}
\alias{alignStreams}
\alias{MultiRatePhysioExperiment}
\alias{experiments}
\alias{experiments<-}
\alias{modalities}
\alias{samplingRates}
\alias{nModalities}
\alias{alignment}
\alias{alignment<-}
\alias{migrateContainer}
\alias{readPhysioRDS}
\alias{assaySamplingRates}
\alias{setAssaySamplingRate}
\alias{PhysioLongitudinal}
\alias{sessions}
\alias{sessions<-}
\alias{session}
\alias{design}
\alias{design<-}
\alias{subjectData}
\alias{subjectData<-}
\alias{addSession}
\alias{asMultiAssayExperiment}
\alias{PhysioCohort}
\alias{subjects}
\alias{subjects<-}
\alias{subject}
\alias{cohortData}
\alias{cohortData<-}
\alias{subjectIds}
\alias{nSubjects}
\alias{groups}
\alias{addSubject}
\alias{subsetCohort}
\alias{cohortDesign}
\alias{aggregatePhysioFeatures}
\alias{harmonizeChannels}
\alias{harmonizeReference}
\alias{harmonizeMontage}
\alias{harmonize}
\alias{harmonizeReport}
\alias{changeScores}
\alias{asMSKTracker}
\alias{AnalysisResult}
\alias{PhysioBiomarker}
\alias{physioBiomarker}
\alias{is.PhysioBiomarker}
\alias{resultType}
\alias{resultValue}
\alias{biomarkerValue}
\alias{estimateOf}
\alias{uncertaintyOf}
\alias{provenanceOf}
\alias{estimandOf}
\alias{NormativeReference}
\alias{normativeLookup}
\alias{zScore}
\alias{percentPredicted}
\alias{icc}
\alias{sem}
\alias{mdc}
\alias{cohensD}
\alias{etaSquared}
\alias{blandAltman}
\alias{circularSummary}
\alias{rayleighTest}
\alias{watsonWilliamsTest}
\alias{circularLinearCorrelation}
\alias{waveformCMC}
\alias{waveformICC}
\alias{waveformReliability}
\alias{fPCA}
\alias{reconstructFPCA}
\alias{registerCurves}
\alias{physioPalette}
\alias{theme_physio}
\alias{scale_color_physio}
\alias{scale_colour_physio}
\alias{scale_fill_physio}
\alias{provenance}
\alias{provenance<-}
\alias{appendProvenance}
\alias{logStep}
\alias{addProvenance}
\alias{withProvenance}
\alias{provenanceHash}
\alias{recordActivity}
\alias{samplingRate}
\alias{samplingRate<-}
\alias{defaultAssay}
\alias{timeIndex}
\alias{duration}
\alias{wpliEstimate}
\alias{.fitMVAR}
\alias{mvarOrderSelect}
\alias{.mvarSpectral}
\alias{mvarFit}
\alias{mvarTransfer}
\alias{channelInfo}
\alias{channelInfo<-}
\alias{channelNames}
\alias{channelNames<-}
\alias{nChannels}
\alias{setChannelTypes}
\alias{setChannelUnits}
\alias{getChannelsByType}
\alias{pickChannels}
\alias{dropChannels}
\alias{renameChannels}
\alias{setReference}
\alias{getReference}
\alias{setElectrodePositions}
\alias{getElectrodePositions}
\alias{applyMontage}
\alias{getEvents}
\alias{setEvents}
\alias{addEvents}
\alias{removeEvents}
\alias{nEvents}
\alias{timeToSamples}
\alias{samplesToTime}
\alias{eventQuery}
\alias{filterType}
\alias{filterValue}
\alias{resolveQuery}
\alias{cbindPhysio}
\alias{rbindPhysio}
\alias{extractWindow}
\alias{checkNA}
\alias{handleNA}
\alias{fillEdgeNA}
\alias{hasNA}
\alias{naSummary}
\alias{replaceNA}
\alias{deprecate_physio}
\alias{registerReader}
\alias{getReader}
\alias{availableReaders}
\alias{unregisterReader}
\alias{registerWriter}
\alias{getWriter}
\alias{availableWriters}
\alias{unregisterWriter}
\alias{registerOperation}
\alias{getOperation}
\alias{availableOperations}
\alias{unregisterOperation}
\alias{EventQuery}
\alias{PhysioExperiment-class}
\alias{PhysioEvents-class}
\alias{MultiPhysioExperiment-class}
\alias{MultiRatePhysioExperiment-class}
\alias{PhysioLongitudinal-class}
\alias{PhysioCohort-class}
\alias{AnalysisResult-class}
\alias{PhysioBiomarker-class}
\alias{NormativeReference-class}
\alias{EventQuery-class}
\alias{AnalysisResult-accessors}
\alias{resultType,AnalysisResult-method}
\alias{resultValue,AnalysisResult-method}
\alias{resultValue,PhysioBiomarker-method}
\alias{biomarkerValue,PhysioBiomarker-method}
\alias{AnalysisResult-carrier}
\alias{estimateOf,AnalysisResult-method}
\alias{uncertaintyOf,AnalysisResult-method}
\alias{provenanceOf,AnalysisResult-method}
\alias{estimandOf,AnalysisResult-method}
\alias{show,AnalysisResult-method}
\alias{as.data.frame,PhysioExperiment-method}
\alias{as.data.frame.PhysioBiomarker}
\alias{cohortData,PhysioCohort-method}
\alias{cohortData<-,PhysioCohort-method}
\alias{design,PhysioLongitudinal-method}
\alias{design<-,PhysioLongitudinal-method}
\alias{dim,PhysioExperiment-method}
\alias{.is_legacy_container}
\alias{.measured_times}
\alias{.onLoad}
\alias{format.PhysioBiomarker}
\alias{length,PhysioExperiment-method}
\alias{updateObject,MultiPhysioExperiment-method}
\alias{multiphysio-legacy-accessors}
\alias{experiments,MultiPhysioExperiment-method}
\alias{experiments<-,MultiPhysioExperiment-method}
\alias{modalities,MultiPhysioExperiment-method}
\alias{samplingRates,MultiPhysioExperiment-method}
\alias{nModalities,MultiPhysioExperiment-method}
\alias{alignment,MultiPhysioExperiment-method}
\alias{alignment<-,MultiPhysioExperiment-method}
\alias{MultiPhysioExperiment-methods}
\alias{length,MultiPhysioExperiment-method}
\alias{names,MultiPhysioExperiment-method}
\alias{[[,MultiPhysioExperiment,ANY,ANY-method}
\alias{dim,MultiPhysioExperiment-method}
\alias{show,MultiPhysioExperiment-method}
\alias{na-handling}
\alias{show,NormativeReference-method}
\alias{NormativeReference-compare}
\alias{zScore,NormativeReference-method}
\alias{percentPredicted,NormativeReference-method}
\alias{physio-registry}
\alias{show,PhysioBiomarker-method}
\alias{PhysioCohort-methods}
\alias{length,PhysioCohort-method}
\alias{[,PhysioCohort,ANY,ANY,ANY-method}
\alias{show,PhysioCohort-method}
\alias{PhysioLongitudinal-methods}
\alias{length,PhysioLongitudinal-method}
\alias{[,PhysioLongitudinal,ANY,ANY,ANY-method}
\alias{show,PhysioLongitudinal-method}
\alias{print.fpca_result}
\alias{provenance,PhysioExperiment-method}
\alias{provenance<-,PhysioExperiment-method}
\alias{provenance,MultiPhysioExperiment-method}
\alias{provenance,PhysioLongitudinal-method}
\alias{provenance,PhysioCohort-method}
\alias{samplingRate,PhysioExperiment-method}
\alias{samplingRate<-,PhysioExperiment-method}
\alias{sessions,PhysioLongitudinal-method}
\alias{sessions<-,PhysioLongitudinal-method}
\alias{show,PhysioEvents-method}
\alias{show,PhysioExperiment-method}
\alias{streams,MultiPhysioExperiment-method}
\alias{streams<-,MultiPhysioExperiment-method}
\alias{[,MultiPhysioExperiment,ANY,ANY,ANY-method}
\alias{[,PhysioExperiment,ANY,ANY,ANY-method}
\alias{subjectData,PhysioLongitudinal-method}
\alias{subjectData<-,PhysioLongitudinal-method}
\alias{subjects,PhysioCohort-method}
\alias{subjects<-,PhysioCohort-method}
\alias{summary,PhysioExperiment-method}
\alias{as.data.frame-PhysioExperiment-method}
\alias{channelInfo-set}
\alias{channelNames-set}
\alias{dim-PhysioExperiment-method}
\alias{dot-fitMVAR}
\alias{dot-is_legacy_container}
\alias{dot-measured_times}
\alias{dot-mvarSpectral}
\alias{dot-onLoad}
\alias{length-PhysioExperiment-method}
\alias{physioBiomarker-constructor}
\alias{show-PhysioEvents-method}
\alias{show-PhysioExperiment-method}
\alias{sub-MultiPhysioExperiment-ANY-ANY-ANY-method}
\alias{sub-PhysioExperiment-ANY-ANY-ANY-method}
\alias{summary-PhysioExperiment-method}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from \pkg{PhysioExperiment} and re-exported here
so that code written against \pkg{PhysioCore} keeps working. See the
\pkg{PhysioExperiment} documentation for their behaviour.
}
\examples{
# Re-exported unchanged from PhysioExperiment, so old PhysioCore code runs:
sig <- matrix(rnorm(400 * 2), nrow = 400, ncol = 2)
pe <- PhysioExperiment(assays = list(raw = sig), samplingRate = 200)
samplingRate(pe)
nChannels(pe)
}
