% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiRatePhysioExperiment-class.R
\name{streams}
\alias{streams}
\alias{streams,MultiRatePhysioExperiment-method}
\alias{streams<-}
\alias{streams<-,MultiRatePhysioExperiment-method}
\title{Access the streams of a MultiRatePhysioExperiment}
\usage{
streams(x)

\S4method{streams}{MultiRatePhysioExperiment}(x)

streams(x) <- value

\S4method{streams}{MultiRatePhysioExperiment}(x) <- value
}
\arguments{
\item{x}{A \code{MultiRatePhysioExperiment}.}

\item{value}{A named list / SimpleList of \code{PhysioExperiment} streams.}
}
\value{
\code{streams()} a \code{SimpleList}; setter returns the updated object.
}
\description{
Access the streams of a MultiRatePhysioExperiment
}
